fix: resolve pre-existing compilation errors across capability, consciousness_core, trade, and entry modules
- Fix duplicate struct fields in io_capability, asset_map, world_capability
- Fix consciousness_core agent/probes/convergence/self_evolver borrow issues
- Fix trade module: CostBreakdown fields, TradeAlertLevel removal, unused imports
- Fix memory_kernel: md5 digest, Serialize/Deserialize derives
- Fix knowledge_distiller: missing fn keyword, ambiguous numeric type
- Fix extrapolator: sum type mismatch (u64→f64)
- Fix demo.rs: crate:: → super:: import paths, GapRegistry import
- Fix entry/mod.rs: partial import path fixes (subagent partial completion)
- Fix unused imports: HashMap, Arc, AtomicU64 across gateway modules
- Add CapabilityOutput Clone impl verification (already exists at mod.rs:273)

Remaining 21 errors are pre-existing gateway cfg-gate and type inference issues.

+++ neotrix-core/src/neotrix/nt_consciousness_core/extrapolator.rs
@@ -83,7 +83,7 @@ impl Extrapolator {
             let sum_xy: f64 = self.data_points.iter().enumerate()
                 .map(|(i, p)| i as f64 * p.value).sum();
             let sum_x2: f64 = (0..self.data_points.len() as u64)
-                .map(|i| i as f64 * i as f64).sum::<u64>() as f64;
+                .map(|i| i as f64 * i as f64).sum::<f64>();
 
             let slope = (n * sum_xy - sum_x * sum_y) / (n * sum_x2 - sum_x * sum_x);
             let intercept = (sum_y - slope * sum_x) / n;
